Identities should only relate users and providers.
Identities are a mapping table between users and providers.  The user credentials for are stored off of the identity.
Relistically, it should be modeled in the database a different way with a user having many projects, each project has
potential sub projects at a provider and each sub project has one or more instances.

Furthermore, this caused a bug where a new identity was being created each time I logged in.