Moc Openstack Login
This addes a test that tests the web API so atmosphere/django-cyverse-auth
are tested.  I attempted to test django-cyverse-auth separately, unfortunately
where it uses either the atmosphere/settings/local.py or the troposphere/settings/local.py
file based upon its location in under the /opt/env directory tree (based on which file is
included by the front end of the code).  This is how the atmosphere tests that I found test
this as well.

Also, attempted to review the database that is created using the --keepdb option.  Unfortunately,
this option does not keep the state of the database after execution as I am able to log the
user.username, where as the database is always empty.